Login

+++ Assets/Scripts/Entities/DojoModels/States/CharacterStates/ActionState.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 using Dojo;
 using Dojo.Torii;
 using dojo_bindings;
@@ -9,13 +10,13 @@ public class ActionState : ModelInstance
     private UInt32 match_id;
     private UInt32 character_id;
     private dojo.FieldElement player;
-    private int playerId;
+    private BigInteger playerId;
     private bool action;
     private bool movement;
 
     public uint Match_id { get => match_id; set => match_id = value; }
     public uint Character_id { get => character_id; set => character_id = value; }
-    public int Player_id { get => playerId; set => playerId = value; }
+    public BigInteger Player_id { get => playerId; set => playerId = value; }
     public bool Action { get => action; set => action = value; }
     public bool Movement { get => movement; set => movement = value; }
 
@@ -28,7 +29,7 @@ public class ActionState : ModelInstance
         movement = model.members["movement"].ty.ty_primitive.p_bool;
 
         var playerString = BitConverter.ToString(player.data.ToArray()).Replace("-", "").ToLower();
-        playerId = System.Int32.Parse(playerString, System.Globalization.NumberStyles.AllowHexSpecifier);
+        playerId = BigInteger.Parse(playerString, System.Globalization.NumberStyles.AllowHexSpecifier);
 
         // Debug.Log("ActionState: \n match_id: " + match_id
         //           + "\n character_id: " + character_id
